fix: display hamburger icon
This is a regression due to the absence of the `navbar-dark` class.
When adding it back, the icon is displayed again, but in a grayish color.
It's possible to fix the border-color but not the icon color with pure CSS,
as Bootstrap uses a SVG image with hardcoded color in the data url for `.navbar-toggler-icon`.
There are two possible ways to fix it:
- override the data url to make it use the theme color
- use another icon like the one provided by fontawesome, which will use the color specified in CSS

I went with the second solution.

Fixes #31