Add a check on JQ processing.
JQ processing errors are silently ignored by piped commands in parallel
because they are not the last command of the pipe (curl is).

That could be prevented by using 'set -o pipefail' Bash option but no
way found to set it in parallel command, because of the used shell looks
to not be Bash, hence the option is unknown and provokes an error.