refactor: use shared configuration in CLI
The CLI now supports setting several configurations, so we can now avoid replicating the complete configuration for every environment.
Furthermore, CLI v12 now uses the `production` configuration by default.
The `angular.json` file has been updated to reflect this new simplified behavior.
This should not have an impact on the build scripts or resulting bundles.

In the process, I realized that the upgrade to CLI v12 broke the help pages (that display markdown files from the assets).
It turns out the default behavior changed in the CLI, and we now have to specify to follow symlinks
(all the markdown files point to files outside of the frontend).

An e2e test has been added to make sure these pages work as expected.

I also realized that the complete assets directory was copied,
so the dist directory ended up with all the assets for all the applications in sub-folders.
This is no longer the case.