fix: search descendants
The search descedants checkbox is now enabled on small aggreggation.
It has also been fixed to avoid emitting an event every time an aggregation was selected, which was triggerring a ton of router navigations (and introducing weird behaviors, spotted when adding the Germplasm tab).

Tests have been added to ensure no regressions.