fix(header): proper title for rare app
The component was relying on the title of the aplication which was changed in https://forgemia.inra.fr/urgi-is/data-discovery/-/commit/7c8861f1aac3e9af7c17dbf444d64fede29e848f

It was always displaying the brc4env title since then.

The component now relies on the application name instead, which should be less britle.
A unit test has been added to check the behavior of the component.